Restructure code organization for test package
